add new methods/views to build out many to many relationsips

+++ Library/Models/Book.cs
@@ -133,7 +133,7 @@ namespace Library.Models
       MySqlConnection conn = DB.Connection();
       conn.Open();
       var cmd = conn.CreateCommand() as MySqlCommand;
-      cmd.CommandText = @"DELETE FROM books WHERE id = @bookId;";
+      cmd.CommandText = @"DELETE FROM authors_books WHERE book_id = @bookId; DELETE FROM books WHERE id = @bookId;";
       MySqlParameter bookIdParameter = new MySqlParameter();
       bookIdParameter.ParameterName = "@bookId";
       bookIdParameter.Value = this._bookId;
@@ -167,6 +167,87 @@ namespace Library.Models
       {
         conn.Dispose();
       }
+    }
+
+    public void AddAuthor(Author newAuthor)
+    {
+      MySqlConnection conn = DB.Connection();
+      conn.Open();
+      var cmd = conn.CreateCommand() as MySqlCommand;
+      cmd.CommandText = @"INSERT INTO authors_books (book_id, author_id) VALUES (@bookId, @authorId);";
+      MySqlParameter bookIdParameter = new MySqlParameter();
+      bookIdParameter.ParameterName = "@bookId";
+      bookIdParameter.Value = _bookId;
+      cmd.Parameters.Add(bookIdParameter);
+      MySqlParameter authorIdParameter = new MySqlParameter();
+      authorIdParameter.ParameterName = "@authorId";
+      authorIdParameter.Value = newAuthor.GetAuthorId();
+      cmd.Parameters.Add(authorIdParameter);
+      cmd.ExecuteNonQuery();
+
+      conn.Close();
+      if (conn != null)
+      {
+        conn.Dispose();
+      }
+    }
+
+    public List<Author> GetAuthors()
+    {
+      MySqlConnection conn = DB.Connection();
+      conn.Open();
+      var cmd = conn.CreateCommand() as MySqlCommand;
+      // joining 3 tables authors books authors_books
+      cmd.CommandText = @"SELECT authors.* FROM authors
+        JOIN authors_books ON (authors.id = authors_books.author_id)
+        JOIN books ON (authors_books.book_id = books.id)
+        WHERE books.id = @bookId;";
+      MySqlParameter bookIdParameter = new MySqlParameter();
+      bookIdParameter.ParameterName = "@bookId";
+      bookIdParameter.Value = _bookId;
+      cmd.Parameters.Add(bookIdParameter);
+      MySqlDataReader rdr = cmd.ExecuteReader() as MySqlDataReader;
+      List<Author> authorList = new List<Author>{};
+      while(rdr.Read())
+      {
+        string authorName = rdr.GetString(0);
+        int authorId = rdr.GetInt32(1);
+        Author newAuthor = new Author(authorName, authorId);
+        authorList.Add(newAuthor);
+      }
+      conn.Close();
+      if (conn != null)
+      {
+        conn.Dispose();
+      }
+      return authorList;
+    }
+
+    public static List<Book> GetAvailable()
+    {
+      MySqlConnection conn = DB.Connection();
+      conn.Open();
+      var cmd = conn.CreateCommand() as MySqlCommand;
+      cmd.CommandText = @"SELECT books.* FROM books WHERE book_id IN (SELECT book_id FROM copies WHERE patron_id IS NULL);";
+      List<Book> allBooks = new List<Book>{};
+      var rdr = cmd.ExecuteReader() as MySqlDataReader;
+      int bookId = 0;
+      string bookTitle = "";
+      while (rdr.Read())
+      {
+        bookTitle = rdr.GetString(0);
+        bookId = rdr.GetInt32(1);
+        Book newBook = new Book(bookTitle, bookId);
+        allBooks.Add(newBook);
+      }
+
+      conn.Close();
+      if (conn != null)
+      {
+        conn.Dispose();
+      }
+
+      return allBooks;
     }
   }
 }
